Pagination and Filtering

+++ SecurityLab/Models/Product.cs
@@ -5,7 +5,7 @@ namespace SecurityLab.Models;
 
 public partial class Product
 {
-    public byte ProductId { get; set; }
+    public int ProductId { get; set; }
 
     public string Name { get; set; } = null!;
